+++ js/utils/util.js
@@ -65,6 +65,27 @@ const showErrorMsg = (message) => {
     container.remove();
   }, SHOW_ERROR_TIME);
 };
+// Функция взята из интернета и доработана
+// Источник - https://www.freecodecamp.org/news/javascript-debounce-example
+
+function debounce (callback, timeoutDelay = 500) {
+  // Используем замыкания, чтобы id таймаута у нас навсегда приклеился
+  // к возвращаемой функции с setTimeout, тогда мы его сможем перезаписывать
+  let timeoutId;
+
+  return (...rest) => {
+    // Перед каждым новым вызовом удаляем предыдущий таймаут,
+    // чтобы они не накапливались
+    clearTimeout(timeoutId);
+
+    // Затем устанавливаем новый таймаут с вызовом колбэка на ту же задержку
+    timeoutId = setTimeout(() => callback.apply(this, rest), timeoutDelay);
+
+    // Таким образом цикл «поставить таймаут - удалить таймаут» будет выполняться,
+    // пока действие совершается чаще, чем переданная задержка timeoutDelay
+  };
+}
+
 
 function getRandomPositiveFloat(abc, bcd, digits = 0) {
   // Чтобы не заставлять пользователя нашей функции помнить порядок аргументов,
@@ -87,4 +108,11 @@ function getRandomPositiveFloat(abc, bcd, digits = 0) {
 }
 getRandomPositiveFloat();
 
-export { getRandomInt, showErrorMsg, getRandomElement, randomNumber, debounce, getArray };
+export {
+  getRandomInt,
+  debounce, showErrorMsg,
+  getRandomElement,
+  randomNumber,
+  debounce,
+  getArray,
+};
